Alpha 1.1.0

+++ src/kernel/include/kernel/io.h
@@ -17,11 +17,11 @@
 /* In functions */
 uint8_t   inb (uint16_t port);
 uint16_t  inw (uint16_t port);
-uint32_t  ind (uint32_t port);
+uint32_t  ind (uint16_t port);
 
 /* Out functions */
-void      outb (uint8_t value, uint16_t port);
-void      outw (uint16_t value, uint16_t port);
-void      outd (uint32_t value, uint32_t port);
+void  outb (uint8_t value, uint16_t port);
+void  outw (uint16_t value, uint16_t port);
+void  outd (uint32_t value, uint16_t port);
 
 #endif
